requirements.txt
numpy
pandas
tensorflow
pillow
protobuf
object_detection
opencv-python
protos